+++ base/src/main/java/proguard/evaluation/value/DetailedArrayReferenceValue.java
@@ -139,7 +139,7 @@ public class DetailedArrayReferenceValue extends IdentifiedArrayReferenceValue
             indexValue.isParticular())
         {
             int index = indexValue.value();
-            if (PartialEvaluator.ENABLE_NEW_EXCEPTIONS && (index < 0 || index >= values.length))
+            if (index < 0 || index >= values.length)
             {
                 throw new ArrayIndexOutOfBounds(index, values.length);
             }
@@ -159,7 +159,7 @@ public class DetailedArrayReferenceValue extends IdentifiedArrayReferenceValue
             if (indexValue.isParticular())
             {
                 int index = indexValue.value();
-                if (PartialEvaluator.ENABLE_NEW_EXCEPTIONS && (index < 0 || index >= values.length))
+                if (index < 0 || index >= values.length)
                 {
                     throw new ArrayIndexOutOfBounds(index, values.length);
                 }